+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @reservoir0x/relay-kit-hooks
 
+## 1.9.3
+
+### Patch Changes
+
+- 6940a55: Gas top up functionality
+- Updated dependencies [6940a55]
+  - @reservoir0x/relay-sdk@1.7.0
+
+## 1.9.2
+
+### Patch Changes
+
+- 293486b: Add additional refund reason messaging
+- Updated dependencies [293486b]
+  - @reservoir0x/relay-sdk@1.6.15
+
 ## 1.9.1
 
 ### Patch Changes